// src/enum.ts
export enum orderRole {
  Pending = "Pending",
  Preparing = "Preparing",
  Ready = "Ready",
  Delivered = "Delivered",
}

export enum paymentRole {
  Paid = "Paid",
  Pending="Pending",
  Failed = "Failed",
}
export enum reservationRole {
  Confirmed = "confirmed",
  Canceled = "canceled",
  Completed = "completed",
}
export enum UserRole {
  Admin = "Admin",
  Employee = "Employee",
  Customer = "Customer",
}